refactor: 댓글 및 대댓글 삭제 전 confirm 확인하는 로직 추가

+++ src/pages/FeedDetail/components/Comment.tsx
@@ -60,7 +60,8 @@ const Comment = ({
   };
 
   const onDeleteComment = () => {
-    deleteComment(parentCommentId);
+    //TODO: #54 머지 후 Confirm 컴포넌트로 대체
+    if (confirm('댓글을 삭제하시겠습니까?')) deleteComment(parentCommentId);
   };
 
   useEffect(() => {
